OTA-2488: Initial move of the virtual secondary implementation out of libaktualizr
