+++ src/libaktualizr/primary/CMakeLists.txt
@@ -12,9 +12,12 @@ set(HEADERS aktualizr.h
 
 
 add_library(primary OBJECT ${SOURCES})
+target_include_directories(primary PUBLIC ${PROJECT_SOURCE_DIR}/src/virtual_secondary)
 
 add_aktualizr_test(NAME aktualizr SOURCES aktualizr_test.cc PROJECT_WORKING_DIRECTORY ARGS ${PROJECT_BINARY_DIR}/uptane_repos)
 add_dependencies(t_aktualizr uptane_repo_full_no_correlation_id)
+target_include_directories(t_aktualizr PUBLIC ${PROJECT_SOURCE_DIR}/src/virtual_secondary)
+target_link_libraries(t_aktualizr virtual-secondary)
 
 if (BUILD_OSTREE)
     add_aktualizr_test(NAME aktualizr_fullostree SOURCES aktualizr_fullostree_test.cc PROJECT_WORKING_DIRECTORY ARGS $<TARGET_FILE:aktualizr-repo> ${PROJECT_BINARY_DIR}/ostree_repo)
@@ -31,5 +34,9 @@ add_aktualizr_test(NAME emptytargets SOURCES empty_targets_test.cc PROJECT_WORKI
 add_aktualizr_test(NAME device_cred_prov SOURCES device_cred_prov_test.cc PROJECT_WORKING_DIRECTORY)
 set_tests_properties(test_device_cred_prov PROPERTIES LABELS "crypto")
 
+add_aktualizr_test(NAME uptane_key SOURCES uptane_key_test.cc PROJECT_WORKING_DIRECTORY)
+set_tests_properties(test_uptane_key PROPERTIES LABELS "crypto")
+target_include_directories(t_uptane_key PUBLIC ${PROJECT_SOURCE_DIR}/src/virtual_secondary)
+target_link_libraries(t_uptane_key virtual-secondary)
 
 aktualizr_source_file_checks(${SOURCES} ${HEADERS} ${TEST_SOURCES})
