+++ src/virtual_secondary/managedsecondary.h
